give help ux, share fixes

+++ deploy-website.sh
@@ -59,17 +59,20 @@ cd ../../
 # Run gulp
 gulp deploy --debug --production
 
-# Move to created directory
-cd _dist
-
-# Push to git by overriding previous commits
-# IMPORTANT: Supress messages so nothing appears in logs
-if [[ $TRAVIS_BRANCH == 'release' ]] || [[ $TRAVIS_BRANCH == 'uat' ]] || [[ $TRAVIS_BRANCH == 'develop' ]]
+if [[ $TRAVIS_PULL_REQUEST == 'false' ]]
   then
-    git init
-    git add -A
-    git commit -m "Travis CI automatic build for $THE_COMMIT"
-    git push --quiet --force "https://${AZURE_USER}:${AZURE_PASSWORD}@${AZURE_WEBSITE}.scm.azurewebsites.net:443/${AZURE_WEBSITE}.git" master > /dev/null 2>&1
-  else
-    echo "Not on a build branch so don't push the changes"
+    # Move to created directory
+    cd _dist
+
+    # Push to git by overriding previous commits
+    # IMPORTANT: Supress messages so nothing appears in logs
+    if [[ $TRAVIS_BRANCH == 'release' ]] || [[ $TRAVIS_BRANCH == 'uat' ]] || [[ $TRAVIS_BRANCH == 'develop' ]]
+      then
+        git init
+        git add -A
+        git commit -m "Travis CI automatic build for $THE_COMMIT"
+        git push --quiet --force "https://${AZURE_USER}:${AZURE_PASSWORD}@${AZURE_WEBSITE}.scm.azurewebsites.net:443/${AZURE_WEBSITE}.git" master > /dev/null 2>&1
+      else
+        echo "Not on a build branch so don't push the changes"
+    fi
 fi
